requirements.txt
--extra-index-url https://pypi.org/simple
python-socketio
fastapi

### Server
gunicorn
eventlet
uvicorn[standard]